refactor: 배럴 파일 생성 및 적용 (#15)
