+++ src/shared/ui/index.ts
@@ -1,0 +1,6 @@
+import { Toaster } from '@/shared/ui/Toaster'
+import Footer from '@/shared/ui/footer/Footer'
+import { Modal, ModalClose } from '@/shared/ui/Modal'
+import Header from '@/shared/ui/header/Header'
+
+export { Toaster, Footer, Modal, ModalClose, Header }
